ethclient: move test
